UPDATE_PERSON permission check added

+++ src/main/java/com/zuehlke/securesoftwaredevelopment/controller/PersonsController.java
@@ -1,9 +1,11 @@
 package com.zuehlke.securesoftwaredevelopment.controller;
 
 import com.zuehlke.securesoftwaredevelopment.config.AuditLogger;
+import com.zuehlke.securesoftwaredevelopment.config.SecurityUtil;
 import com.zuehlke.securesoftwaredevelopment.domain.Person;
 import com.zuehlke.securesoftwaredevelopment.domain.User;
 import com.zuehlke.securesoftwaredevelopment.repository.PersonRepository;
+import com.zuehlke.securesoftwaredevelopment.repository.RoleRepository;
 import com.zuehlke.securesoftwaredevelopment.repository.UserRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -18,6 +20,7 @@ import javax.servlet.http.HttpSession;
 import java.nio.file.AccessDeniedException;
 import java.sql.SQLException;
 import java.util.List;
+import java.util.Objects;
 
 @Controller
 
@@ -28,15 +31,24 @@ public class PersonsController {
 
     private final PersonRepository personRepository;
     private final UserRepository userRepository;
+    private final RoleRepository roleRepository;
 
-    public PersonsController(PersonRepository personRepository, UserRepository userRepository) {
+    public PersonsController(PersonRepository personRepository, UserRepository userRepository, RoleRepository roleRepository) {
         this.personRepository = personRepository;
         this.userRepository = userRepository;
+        this.roleRepository = roleRepository;
     }
 
     @GetMapping("/persons/{id}")
-    @PreAuthorize("hasAuthority('VIEW_PERSON')")
-    public String person(@PathVariable int id, Model model, HttpSession session) {
+    @PreAuthorize("hasAuthority('VIEW_PERSON') || hasAuthority('UPDATE_PERSON')")
+    public String person(@PathVariable int id, Model model, HttpSession session) throws AccessDeniedException{
+
+        int currentUserId = Objects.requireNonNull(SecurityUtil.getCurrentUser()).getId();
+        boolean currentUserIsNotAdmin = roleRepository.findByUserId(currentUserId)
+                .stream().noneMatch(role -> role.getName().equals("ADMIN"));
+
+        if(currentUserIsNotAdmin && currentUserId != id) throw new AccessDeniedException("Access denied!");
+
         String csrfToken = session.getAttribute("CSRF_TOKEN").toString();
         model.addAttribute("CSRF_TOKEN", csrfToken);
         model.addAttribute("person", personRepository.get("" + id));
@@ -44,14 +56,24 @@ public class PersonsController {
     }
 
     @GetMapping("/myprofile")
-    public String self(Model model, Authentication authentication) {
+    public String self(Model model, Authentication authentication, HttpSession session) {
+        String csrfToken = session.getAttribute("CSRF_TOKEN").toString();
+        model.addAttribute("CSRF_TOKEN", csrfToken);
         User user = (User) authentication.getPrincipal();
         model.addAttribute("person", personRepository.get("" + user.getId()));
         return "person";
     }
 
     @DeleteMapping("/persons/{id}")
-    public ResponseEntity<Void> person(@PathVariable int id) {
+    @PreAuthorize("hasAuthority('UPDATE_PERSON')")
+    public ResponseEntity<Void> person(@PathVariable int id) throws AccessDeniedException  {
+
+        int currentUserId = Objects.requireNonNull(SecurityUtil.getCurrentUser()).getId();
+        boolean currentUserIsNotAdmin = roleRepository.findByUserId(currentUserId)
+                .stream().noneMatch(role -> role.getName().equals("ADMIN"));
+
+        if(currentUserIsNotAdmin && currentUserId != id) throw new AccessDeniedException("Access denied!");
+
         personRepository.delete(id);
         userRepository.delete(id);
 
@@ -59,11 +81,18 @@ public class PersonsController {
     }
 
     @PostMapping("/update-person")
+    @PreAuthorize("hasAuthority('UPDATE_PERSON')")
     public String updatePerson(Person person, HttpSession session, @RequestParam("csrfToken") String csrfToken) throws AccessDeniedException {
 
         String sessionToken = session.getAttribute("CSRF_TOKEN").toString();
 
         if(!csrfToken.equals(sessionToken)) throw new AccessDeniedException("Access denied!");
+
+        int currentUserId = Objects.requireNonNull(SecurityUtil.getCurrentUser()).getId();
+        boolean currentUserIsNotAdmin = roleRepository.findByUserId(currentUserId)
+                .stream().noneMatch(role -> role.getName().equals("ADMIN"));
+
+        if(currentUserIsNotAdmin && currentUserId != Integer.parseInt(person.getId())) throw new AccessDeniedException("Access denied!");
 
         personRepository.update(person);
